+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,1 +1,1 @@
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.6-bin.zip
+distributionUrl=https\://mirrors.cloud.tencent.com/gradle/gradle-8.2.1-all.zip
