--- src/main/java/com/mixin/Jade.java
@@ -1,11 +1,0 @@
-package com.mixin;
-
-public class Jade {
-
-
-
-
-
-
-
-}
