+++ src/main/java/minearchive/minearchive/Config.java
@@ -1,0 +1,11 @@
+package minearchive.minearchive;
+
+import minearchive.Minearchive;
+import net.minecraftforge.common.ForgeConfigSpec;
+import net.minecraftforge.fml.common.Mod;
+
+@Mod.EventBusSubscriber(modid = Minearchive.MODID, bus = Mod.EventBusSubscriber.Bus.MOD)
+public class Config {
+    private static final ForgeConfigSpec.Builder BUILDER = new ForgeConfigSpec.Builder();
+    static final ForgeConfigSpec SPEC = BUILDER.build();
+}
